requirements/develop.txt
# Include all project modules
-r project.txt

django-extensions
werkzeug
